NIAD-910: Refactoring logic to separate mappers.
