+++ service/src/main/java/uk/nhs/adaptors/gp2gp/ehr/mapper/CommentType.java
@@ -1,0 +1,13 @@
+package uk.nhs.adaptors.gp2gp.ehr.mapper;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public enum CommentType {
+    LABORATORY_RESULT_COMMENT("LABORATORY RESULT COMMENT(E141)"),
+    USER_COMMENT("USER COMMENT");
+
+    private final String code;
+}
