+++ service/src/main/resources/templates/observation_narrative_statement_template.mustache
@@ -6,7 +6,7 @@
 
             {{comment}}</text>
         <statusCode code="COMPLETE"/>
-        <availabilityTime value="{{issuedDate}}"/>
+        {{{availabilityTime}}}
         {{#participant}}
             {{{.}}}
         {{/participant}}
