+++ service/src/main/resources/templates/observation_statement_template.mustache
@@ -6,7 +6,7 @@
         <effectiveTime>
             {{{effectiveTime}}}
         </effectiveTime>
-        <availabilityTime value="{{issuedDate}}"/>
+        {{{availabilityTime}}}
         {{{value}}}
         {{{referenceRange}}}
         {{{interpretation}}}
